udp: widen sendMany loop counter from u16 to u32
JSArrayIterator uses a u32 internally and array_len is a usize, so a
connected socket with more than 65535 payloads (or an unconnected
socket with more than 196608 triple-encoded elements) overflowed the
u16 counter on increment. Pre-existing issue in the same loop that the
connect_info snapshot fix touches.

+++ src/bun.js/api/bun/udp_socket.zig
@@ -636,7 +636,7 @@ pub const UDPSocket = struct {
 
         var iter = try arg.arrayIterator(globalThis);
 
-        var i: u16 = 0;
+        var i: u32 = 0;
         var port: JSValue = .zero;
         while (try iter.next()) |val| : (i += 1) {
             if (i >= array_len) {
